+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Switched to a BTreeMap instead of Vec to get automatic deduplication of entries
 * Updated Rusoto to 0.42
+* Check for destination bucket access before running the analysis
 
 ### 0.2.0 - 11/15/2019